fix influx export printer live data
