+++ server/cache/influx-export.cache.js
@@ -1,0 +1,16 @@
+const InfluxCleanerService = require("../services/influx-cleaner.service");
+
+let influxCache = undefined;
+
+function getInfluxCleanerCache() {
+  if (!!influxCache) {
+    return influxCache;
+  } else {
+    influxCache = new InfluxCleanerService();
+    return influxCache;
+  }
+}
+
+module.exports = {
+  getInfluxCleanerCache
+};
