+++ server/routes/sse.printer-monitoring.routes.js
@@ -15,13 +15,11 @@ let interval = false;
 
 const { SettingsClean } = require("../services/settings-cleaner.service.js");
 const { getSorting, getFilter } = require("../services/front-end-sorting.service.js");
-const { writePoints } = require("../services/influx-export.service.js");
+const { getInfluxCleanerCache } = require("../cache/influx-export.cache");
 // User Modal
 const { ensureCurrentUserAndGroup } = require("../middleware/users.js");
 const { getPrinterStoreCache } = require("../cache/printer-store.cache");
 const { getPrinterManagerCache } = require("../cache/printer-manager.cache");
-
-let influxCounter = 5000;
 
 const sortMe = function (printers) {
   const sortBy = getSorting();
@@ -95,6 +93,12 @@ const filterMe = function (printers) {
   }
 };
 async function sendData() {
+  try {
+    getInfluxCleanerCache().cleanAndWritePrintersInformationForInflux();
+  } catch (e) {
+    logger.error("Unable to clean and write information to influx database!", e.toString());
+  }
+
   const currentOperations = getCurrentOperations();
 
   let printersInformation = getPrinterStoreCache().listPrintersInformationForMonitoringViews();
@@ -102,22 +106,6 @@ async function sendData() {
   printersInformation = await filterMe(printersInformation);
   printersInformation = sortMe(printersInformation);
   const printerControlList = getPrinterManagerCache().getPrinterControlList();
-
-  let serverSettings = SettingsClean.returnSystemSettings();
-  if (typeof serverSettings === "undefined") {
-    await SettingsClean.start();
-    serverSettings = SettingsClean.returnSystemSettings();
-  }
-
-  if (!!serverSettings.influxExport?.active) {
-    if (influxCounter >= 5000) {
-      sendToInflux(printersInformation);
-      influxCounter = 0;
-    } else {
-      influxCounter = influxCounter + 500;
-    }
-    // eslint-disable-next-line no-use-before-define
-  }
 
   for (clientId in clients) {
     let clientSettings = SettingsClean.returnClientSettings(
@@ -162,143 +150,10 @@ router.get("/get/", ensureAuthenticated, ensureCurrentUserAndGroup, async functi
   await sendData();
 });
 
-function sendToInflux(printersInformation) {
-  printersInformation.forEach((printer) => {
-    if (printer.printerState.colour.category !== "Offline") {
-      const date = Date.now();
-
-      const tags = {
-        name: printer?.printerName ? printer.printerName : " ",
-        group: printer?.group ? printer.group : " ",
-        url: printer?.printerURL ? printer.printerURL : " ",
-        state: printer?.printerState?.state ? printer.printerState.state : " ",
-        stateCategory: printer?.printerState?.colour?.category
-          ? printer.printerState?.colour?.category
-          : " ",
-        host_state: printer?.hostState?.state ? printer.hostState?.state : " ",
-        websocket_state: printer?.webSocketState?.colour ? printer.webSocketState?.colour : " ",
-        octoprint_version: printer?.octoPrintVersion ? printer.octoPrintVersion : " "
-      };
-      const printerData = {
-        name: printer?.printerName ? printer.printerName : " ",
-        group: printer?.group ? printer.group : " ",
-        url: printer?.printerURL ? printer.printerURL : " ",
-        state: printer?.printerState?.state ? printer.printerState.state : " ",
-        host_state: printer?.hostState?.state ? printer.hostState.state : " ",
-        websocket_state: printer?.webSocketState?.colour ? printer.webSocketState.colour : " ",
-        octoprint_version: printer?.octoPrintVersion ? printer.octoPrintVersion : " ",
-        state_category: printer?.printerState?.colour?.category
-          ? printer.printerState.colour.category
-          : " ",
-        current_idle_time: printer.currentIdle ? parseFloat(printer.currentIdle) : 0,
-        current_active_time: printer.currentActive ? parseFloat(printer.currentActive) : 0,
-        current_offline_time: printer.currentOffline ? parseFloat(printer.currentOffline) : 0,
-        date_added: printer.dateAdded ? parseFloat(printer.dateAdded) : 0,
-        storage_free: printer?.storage?.free ? parseFloat(printer.storage.free) : 0,
-        storage_total: printer?.storage?.total ? parseFloat(printer.storage.total) : 0,
-        timestamp: date
-      };
-      if (typeof printer.resends !== "undefined") {
-        printerData["job_resends"] = `${printer.resends.count} / ${
-          printer.resends.transmitted / 1000
-        }K (${printer.resends.ratio.toFixed(0)}`;
-      }
-
-      if (typeof printer.currentJob !== "undefined" && printer.currentJob !== null) {
-        for (const key in printer.currentJob) {
-          if (printer.currentJob.hasOwnProperty(key)) {
-            if (key === "progress" && printer.currentJob[key] !== null) {
-              printerData["job_progress"] = parseFloat(printer.currentJob[key]);
-            }
-            if (key === "fileName" && printer.currentJob[key] !== null) {
-              printerData["job_file_name"] = printer.currentJob[key];
-            }
-            if (key === "fileDisplay" && printer.currentJob[key] !== null) {
-              printerData["job_file_display"] = printer.currentJob[key];
-            }
-            if (key === "filePath" && printer.currentJob[key] !== null) {
-              printerData["job_file_path"] = printer.currentJob[key];
-            }
-            if (key === "expectedCompletionDate" && printer.currentJob[key] !== null) {
-              printerData["job_expected_completion_date"] = printer.currentJob[key];
-            }
-            if (key === "expectedPrintTime" && printer.currentJob[key] !== null) {
-              printerData["job_expected_print_time"] = parseFloat(printer.currentJob[key]);
-            }
-            if (key === "expectedFilamentCosts" && printer.currentJob[key] !== null) {
-            }
-            if (key === "expectedPrinterCosts" && printer.currentJob[key] !== null) {
-              printerData["job_expected_print_cost"] = parseFloat(printer.currentJob[key]);
-            }
-            if (key === "expectedTotals" && printer.currentJob[key] !== null) {
-            }
-            if (key === "currentZ" && printer.currentJob[key] !== null) {
-              printerData["job_current_z"] = parseFloat(printer.currentJob[key]);
-            }
-            if (key === "printTimeElapsed" && printer.currentJob[key] !== null) {
-              printerData["job_print_time_elapsed"] = parseFloat(printer.currentJob[key]);
-            }
-            if (key === "printTimeRemaining" && printer.currentJob[key] !== null) {
-              printerData["job_print_time_remaining"] = parseFloat(printer.currentJob[key]);
-            }
-            if (key === "averagePrintTime" && printer.currentJob[key] !== null) {
-              printerData["job_average_print_time"] = parseFloat(printer.currentJob[key]);
-            }
-            if (key === "lastPrintTime" && printer.currentJob[key] !== null) {
-              printerData["job_last_print_time"] = parseFloat(printer.currentJob[key]);
-            }
-            if (key === "thumbnail" && printer.currentJob[key] !== null) {
-              printerData["job_thumbnail"] = printer.currentJob[key];
-            }
-          }
-        }
-      }
-
-      if (printer.selectedFilament.length >= 1) {
-        printer.selectedFilament.forEach((spool, index) => {
-          if (spool !== null) {
-            printerData[`tool_${index}_spool_name`] = spool.spools.name;
-            printerData[`tool_${index}_spool_used`] = parseFloat(spool.spools.used);
-            printerData[`tool_${index}_spool_weight`] = parseFloat(spool.spools.weight);
-            printerData[`tool_${index}_spool_temp_offset`] = parseFloat(spool.spools.tempOffset);
-            if (typeof spool.spools.material !== "undefined") {
-              printerData[`tool_${index}_spool_material`] = spool.spools.material;
-            }
-          }
-        });
-      }
-
-      if (
-        typeof printer.tools !== "undefined" &&
-        printer.tools !== null &&
-        printer.tools[0] !== null
-      ) {
-        for (const key in printer.tools[0]) {
-          if (printer.tools[0].hasOwnProperty(key)) {
-            if (key !== "time") {
-              if (printer.tools[0][key].actual !== null) {
-                printerData[key + "_actual"] = parseFloat(printer.tools[0][key].actual);
-              } else {
-                printerData[key + "_actual"] = 0;
-              }
-              if (printer.tools[0][key].target !== null) {
-                printerData[key + "_target"] = parseFloat(printer.tools[0][key].target);
-              } else {
-                printerData[key + "_target"] = 0;
-              }
-            }
-          }
-        }
-      }
-      writePoints(tags, "PrintersInformation", printerData);
-    }
-  });
-}
-
 if (interval === false) {
   interval = setInterval(async function () {
     await sendData();
-  }, 500);
+  }, 1000);
 }
 
 module.exports = router;
