+++ server/tasks.js
@@ -13,6 +13,7 @@ const {
   updatePluginNoticesStore,
   updatePluginStore
 } = require("./store/octoprint-plugin-list.store");
+const { getInfluxCleanerCache } = require("./cache/influx-export.cache");
 const { FileClean } = require("./services/file-cleaner.service");
 const { sortCurrentOperations } = require("./services/printer-statistics.service");
 const { initFarmInformation } = require("./services/farm-information.service");
@@ -30,20 +31,20 @@ const I_AM_ALIVE = () => {
   });
 };
 
+const INITIALISE_SYSTEM_CACHE = async () => {
+  await getPrinterManagerCache();
+  await getPrinterStoreCache();
+  await getEventEmitterCache();
+  await initHistoryCache();
+  await getInfluxCleanerCache();
+};
+
 const INITIALISE_PRINTERS = async () => {
   await getPrinterManagerCache().initialisePrinters();
 };
 
 const START_PRINTER_ADD_QUEUE = async () => {
   await getPrinterManagerCache().startPrinterEnableQueue();
-};
-
-const INITIALIST_PRINTERS_STORE = async () => {
-  await getPrinterStoreCache();
-};
-
-const INITIALIST_EVENT_EMITTER_CACHE = async () => {
-  await getEventEmitterCache();
 };
 
 const SORT_CURRENT_OPERATIONS = async () => {
@@ -57,12 +58,6 @@ const CRASH_TEST_TASK = async () => {
 
 const FARMPI_DETECTION_TASK = async () => {
   await detectFarmPi();
-};
-
-const HISTORY_CACHE_TASK = async () => {
-  await initHistoryCache().catch((e) => {
-    console.error("X HistoryCache failed to initiate. " + e);
-  });
 };
 
 const FILAMENT_CLEAN_TASK = async () => {
@@ -152,6 +147,7 @@ function TaskStart(task, preset, milliseconds = 0) {
 
 class OctoFarmTasks {
   static RECURRING_BOOT_TASKS = [
+    TaskStart(INITIALISE_SYSTEM_CACHE, TaskPresets.RUNONCE),
     TaskStart(SYSTEM_INFO_CHECK_TASK, TaskPresets.RUNONCE),
     TaskStart(GITHUB_UPDATE_CHECK_TASK, TaskPresets.PERIODIC_IMMEDIATE_DAY),
     TaskStart(UPDATE_OCTOPRINT_PLUGINS_LIST, TaskPresets.PERIODIC_IMMEDIATE_DAY),
@@ -160,14 +156,11 @@ class OctoFarmTasks {
     TaskStart(MEMORY_PROFILING_TASK, TaskPresets.PERIODIC_10000MS),
     TaskStart(FARMPI_DETECTION_TASK, TaskPresets.RUNONCE),
     TaskStart(INIT_FARM_INFORMATION, TaskPresets.RUNONCE),
-    TaskStart(INITIALIST_PRINTERS_STORE, TaskPresets.RUNONCE),
     TaskStart(INITIALISE_PRINTERS, TaskPresets.RUNONCE),
     TaskStart(SORT_CURRENT_OPERATIONS, TaskPresets.PERIODIC_1000MS),
     TaskStart(GENERATE_PRINTER_CONTROL_LIST, TaskPresets.PERIODIC_5000MS),
-    TaskStart(INITIALIST_EVENT_EMITTER_CACHE, TaskPresets.RUNONCE),
     TaskStart(STATE_TRACK_COUNTERS, TaskPresets.PERIODIC, 30000),
     TaskStart(FILAMENT_CLEAN_TASK, TaskPresets.RUNDELAYED, 1000),
-    TaskStart(HISTORY_CACHE_TASK, TaskPresets.RUNONCE),
     TaskStart(GENERATE_MONTHLY_HISTORY_STATS, TaskPresets.PERIODIC_IMMEDIATE_DAY),
     TaskStart(RUN_PRINTER_HEALTH_CHECKS, TaskPresets.PERIODIC_600000MS),
     TaskStart(GENERATE_FILE_STATISTICS, TaskPresets.RUNONCE),
